vinyl: implement page cache
Every vinyl's run iterator independently keeps 2 uncompressed pages
stored in memory - the current one and the previous one used by this
exact iterator. 2 most recently requested pages stored for each run
iterator do not provide much of caching effect and are used mainly to
boost working with multi-page data chunks. So in practice if a tuple is
not found in tuple cache, the engine almost always has to read run files.
The latter incurs severe overhead even if the pages being read happen to
be found in system page cache, because run files are compressed and we
have to decompress them on each read. Besides, relying on system page
cache eviction algorithm may result in sub-optimal performance.

To address this problem, we make vinyl store some amount of unused pages
in memory (up to page cache memory quota), giving preference to most
recently requested (among all runs) ones. These pages are stored in
uncompressed state, so in case of page cache hit no reading from the disk
or decompression is required.

As an additional effect, by making the way of storing pages in memory
suitable for page cache usage, we prevent duplicating pages in memory.
Currently, if an iterator receives a request on a page that is already
stored in memory but attached to another iterator, it will re-read this
page from the disk and decompress it, creating a page's copy. We make run
iterators access pages via run-wise centralized metadata structures, so
that no more duplicates of existing (in memory) pages are created.

Closes #3202

@TarantoolBot document
Title: new box.cfg parameter: 'vinyl_page_cache'

`vinyl_page_cache` puts a limit on the amount of RAM that can be occupied
by currently unused runs' pages, stored in Vinyl's page cache.

The `vinyl_page_cache` limit can be periodically exceeded by up to the
approximate size of one page + it's metadata, so that pages are not
deleted from the cache after being loaded from the disk but before being
processed by the engine and pages are not deleted right before being
loaded again upon request.

`vinyl_page_cache` parameter can be changed dynamically. If the amount of
RAM already being used by the page cache is greater than the new quota,
Vinyl will immediately free enough cache's memory to fit it in the new
limit.

If `vinyl_page_cache` is set to 0, page cache keeps only 1 page in memory
(for the entire engine!), so its performance (and the engine's
performance as well) drops significantly. Therefore it is strongly
recommended not to set `vinyl_page_cache` to 0 if RAM resources are not
exhausted.